minor fix 2
