+++ app/src/main/AndroidManifest.xml
@@ -16,22 +16,16 @@
             android:name=".ui.activity.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.Project1">
-
-        </activity>
-        <activity
-            android:name=".ui.activity.OrderTab"
-            android:exported="true"
-            android:theme="@style/Theme.Project1">
-        </activity>
-        <activity
-            android:name=".ui.activity.ReservationTab"
-            android:exported="true"
-            android:theme="@style/Theme.Project1">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
+        </activity>
+        <activity
+            android:name=".ui.activity.OrderTab"
+            android:exported="true"
+            android:theme="@style/Theme.Project1">
         </activity>
     </application>
