+++ app/src/main/java/com/example/project1/data/Tables.kt
@@ -1,9 +1,13 @@
 package com.example.project1.data
 
+import java.util.Date
+
 data class Tables(
-    val id : Int,
-    val tableNumber: Number,
-    val tableSeatNumber: Number,
-    val tablePosition: Number,
-    val tableStatus: Boolean
+    val tables_id : Int,
+    val name: String,
+    val quantity : Int,
+    val location : String,
+    val status : String,
+    val create_at : Date,
+    val update_at : Date
 )
